Add files via upload
This is the code that causes the placeholder not to go away. This was discovered by Kiala the theme plugin, who was blamed by Woocommerce. It causes more problems with other plugins (advanced fields pro) (category tree) and the site to go dark. Site is antiquedutchclockparts.nl  can you please look into it and solve this issue.